Tones of bug fixes.  AddPurchaseOrder and AddSellOrder might be working fully at this point.  Further testing needed but some testing done.  Delete SO/PO still needs to be added.

+++ ResourceMarketDemo/RMDDatabase/CurrencyExchangeRates.sql
@@ -2,7 +2,7 @@
 (
 	[SourceCurrencyId] TINYINT NOT NULL , 
     [DestinationCurrencyId] TINYINT NOT NULL, 
-    [SourceMultiplier] DECIMAL(38, 9) NOT NULL, 
+    [SourceMultiplier] FLOAT(53) NOT NULL, 
     PRIMARY KEY ([SourceCurrencyId], [DestinationCurrencyId]), 
     CONSTRAINT [CK_CurrencyExchangeRates_IdsNotSame] CHECK ([SourceCurrencyId] <> [DestinationCurrencyId]), 
     CONSTRAINT [FK_CurrencyExchangeRates_CurrencyTypes_SourceId] FOREIGN KEY ([SourceCurrencyId]) REFERENCES [CurrencyTypes]([Id]),
